+++ Assets/terrain/Resource.cs
@@ -3,27 +3,27 @@ using System.Collections;
 
 public class Resource : MonoBehaviour {
     
-	public float value = 1000f;
-    public float depletionRate = 0f;
+	public int value = 1000;
+    public int depletionRate = 0;
     
 	void Start(){
 
 	}
     
-	public float collectBlorb(){
+	public int collectBlorb(){
 		if (value > depletionRate){ //if can remove depletionRate amount 
 	        value -= depletionRate;
 	        return depletionRate;
 		}
 		else if (value > 0){
 			Debug.Log ("Case B");
-			float remainder = value;
+			int remainder = value;
 			value = 0;
 			return remainder;
 		}
 		else{
 			Debug.Log ("Case C");
-			return 0f;
+			return 0;
 		}
     }
